refactor: rearrange plugin structure slightly
Since this is not a DIDComm protocol plugin, it doesn't make sense to
follow the same conventions
