--- oid4vci/oid4vci/definition.py
@@ -1,10 +1,0 @@
-"""Version definitions for this plugin."""
-
-versions = [
-    {
-        "major_version": 1,
-        "minimum_minor_version": 0,
-        "current_minor_version": 0,
-        "path": "v1_0",
-    }
-]
